+++ backend/celery_app/tasks/graph.py
@@ -5,7 +5,7 @@ from collections import defaultdict
 from typing import Dict, List
 
 from backend.celery_app.celery import app
-from backend.api.endpoints.graph import _build_graph_data
+from backend.graph.builder import build_graph_data
 from backend.graph.analysis import find_orphan_nodes, get_orphan_degree_threshold
 from backend.schemas.graph import NodeType, GraphNode
 
@@ -22,7 +22,7 @@ def detect_orphan_notes_for_all_users(self):
     logger.info("[%s] 전역 고립 노트 스캔 스케줄러가 시작되었습니다.", task_name)
     
     # 1. 전체 데이터베이스에서 노드와 엣지 빌드 (내부적으로 PARA CATEGORY 등 포함)
-    graph_data = _build_graph_data()
+    graph_data = build_graph_data()
     threshold = get_orphan_degree_threshold()
     
     # 2. 사용자별 노드 그룹화 (보안 필수 요건: hashed_user_id 기반 컨텍스트 주입 및 격리)
@@ -33,8 +33,12 @@ def detect_orphan_notes_for_all_users(self):
             continue
         
         # 보안 장치: PII 마스킹된 hashed_user_id를 기준 키로 사용
-        # 식별되지 않은 파일의 경우 "unassigned" 풀로 격리
-        uid = node.user_id_hash or "unassigned"
+        # 식별되지 않은 파일의 경우 분석 대상에서 제외하여 테넌트 믹스(Data Leakage) 원천 차단
+        if not node.user_id_hash:
+            logger.warning("[%s] user_id_hash가 없는 노드(%s) 발견. 보안 격리를 위해 스캔에서 제외합니다.", task_name, node.id)
+            continue
+            
+        uid = node.user_id_hash
         nodes_by_user[uid].append(node)
         
     total_orphans_found = 0
@@ -50,7 +54,7 @@ def detect_orphan_notes_for_all_users(self):
         user_node_ids = {n.id for n in user_nodes}
         user_edges = [
             e for e in graph_data.edges 
-            if e.source in user_node_ids or e.target in user_node_ids
+            if e.source in user_node_ids and e.target in user_node_ids
         ]
         
         # 해당 사용자 컨텍스트 안에서만 orphan 판별
